+++ engine/public/math/vector2.inl
@@ -1,19 +1,19 @@
 namespace Math
 {
 	inline Vector2::Vector2()
-		: m_x(0.0f)
-		, m_y(0.0f)
+		: m_x(0)
+		, m_y(0)
 	{
 
 	}
 
-	inline Vector2::Vector2(float x, float y)
+	inline Vector2::Vector2(int32_t x, int32_t y)
 		: m_x(x)
 		, m_y(y)
 	{
 	}
 
-	inline Vector2::Vector2(float v)
+	inline Vector2::Vector2(int32_t v)
 		: m_x(v)
 		, m_y(v)
 	{
@@ -31,7 +31,7 @@ namespace Math
 
 	inline Vector2 Vector2::operator*(float f) const
 	{
-		return Vector2(m_x*f, m_y*f);
+		return Vector2((int32_t)(m_x*f), (int32_t)(m_y*f));
 	}
 
 	inline void Vector2::operator*=(float f)
@@ -61,7 +61,7 @@ namespace Math
 
 	inline float Vector2::Length() const
 	{
-		return Math::SquareRoot((m_x*m_x) + (m_y*m_y));
+		return Math::SquareRoot((float)((m_x*m_x) + (m_y*m_y)));
 	}
 
 	inline Vector2 Vector2::Normalised() const
